--- README.md
@@ -1,21 +1,0 @@
-# is1200-pong-project
- Project to make a pong game on a UNO32 chipkit with the basic I/O shield. 
- Part of IS1200 at KTH Royal Institute of Technology. 
- Meets the advanced requirements for the course project.
-
-Requirements:
-- Menu
-- 1 and 2 player modes
-- Difficulty levels for AI
-- AI Higher priority, because of instructions
-
-Highscore list (volatile, min 3 capital initials)
-
-"Player X wins!" screen
-
-Min 1 object of at least 2x2 moving both x and y
-
-Different bounce angles on sides of the padle
-Different bounce angles on padle depending on Y-position
-frame update timer as timer on chipkit
-
