+++ constants.c
@@ -1,0 +1,5 @@
+
+#include "constants.h"
+
+// Define the constant here
+const int BALLSIZE = 2;
