+++ constants.h
@@ -1,0 +1,16 @@
+
+#ifndef CONSTANTS_H
+#define CONSTANTS_H
+
+#define ORIGINAL_WIDTH 128
+#define ORIGINAL_HEIGHT 32
+
+#define SCREEN_SCALING 8
+#define SCREEN_WIDTH (ORIGINAL_WIDTH*SCREEN_SCALING)
+#define SCREEN_HEIGHT (ORIGINAL_HEIGHT*SCREEN_SCALING)
+
+#define FPS 144
+
+extern const int BALLSIZE;
+
+#endif // CONSTANTS_H
