--- gameConstants.h
@@ -1,6 +1,0 @@
-#ifndef GAMECONSTANTS_H
-#define GAMECONSTANTS_H
-
-int defaultPaddleHeight = 8;
-
-#endif
