--- io.h
@@ -1,12 +1,0 @@
-
-#ifndef IO_H
-#define IO_H
-
-int btn3pressed(void);
-int btn2pressed(void);
-int btn1pressed(void);
-int rPressed(void);
-int btn4pressed(void);
-
-#endif
-
