--- mipslabwork.c
@@ -1,23 +1,0 @@
-/* mipslabwork.c
-
-   This file written 2015 by F Lundevall
-   Updated 2017-04-21 by F Lundevall
-
-   This file should be changed by YOU! So you must
-   add comment(s) here with your name(s) and date(s):
-
-   This file modified 2017-04-31 by Ture Teknolog 
-
-   For copyright and licensing, see file COPYING */
-
-#include <stdint.h>   /* Declarations of uint_32 and the like */
-#include "mipslab.h"  /* Declatations for these labs */
-#include <math.h>
-
-/* Lab-specific initialization goes here */
-void pongInit( void )
-{
-  
-
-  /* displaySplashMenu(); */
-}
